requirements.txt
python-telegram-bot==21.4
fastapi==0.112.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
Jinja2==3.1.4
python-multipart==0.0.9
aiofiles==24.1.0
pydantic==2.8.2
requests==2.32.3
PyJWT==2.9.0
alembic==1.13.2
